I went through that road some time ago. In fact, a couple of times.
Well, you do have to write about them. And tourists should make this the second most visited attraction of Granada. Will help towards paying them, for starters.

And this is the first advance in this that will, or will not, become a book in like what? Months? I'd like to finish it some day. Soon. Or not.