Finishing the chapter on whorehouses and such
Which has allowed me to find a true jewel that describes that part of the underground economy in Granada, the paper by my colleague Bautista. It was bound to become a book, and eventually it was just a paper, but it is an incredible source to understand the underbelly of the city where I live.