A bit of snow
I probably had something in mind when I wrote this, so long ago I barely remember. However, it has not snowed for such a long time I can barely remember what to do in Granada when it snows. Besides, it's 7k words or bust today, so it was as good as the next one. Let it snow, then.