Starts talking about geek stuff
Which is something I should be an expert of. I was bound to finish the chapter on the university, which, being $dayjob, should be something I know about, but you have to squash the inspiration when it comes. And push towards 1500 words today, taking it up to a total of 9.5k, which, if it could be 10K, would be past my limit for the whole Christmas and I could devote myself to other more dayjob related endeavors.