Done more than my quota today
It is still a short text, barely a bit more than a magazine article. However, it is better to do a short first edition and keep improving it than delaying it forever until you make a perfect rounded up text. So I'll keep banging on the chapters that are already there, finish the others that have barely started, and release it before the end of Xmastime.