Finishing the Dr. Olóriz story
And bloating even more the geek part of the book, which I will have to spin off eventually. However, how do I do that? Do I make a small synthesis here and just start writing the other? Is it worth the while to do so? The book is not so big anyway, around 100 pages. So let's keep it the way it is, and let's see.