Writing about Emilio Herrera
I know this guy is just incredible, but it is so incredible that I can't believe it all. The wikipedia article has no references, and digging through all of them would take years. It's a lot like he was offered this, but he refused, he did that, but in a short time he was expelled... I'll take it at face value, because if you look just at the stuff he clearly did, he's a great guy.