Remembered about the glacier
And on the way learned a little about it. Diminishing little ole boy...

BTW, sales are doing well. Not incredibly well, but anything more than nothing would be good for me. A dozen books and ebooks have made their way into the hand of people that I possible have never met, which is good, thanks for asking.