how to check out fruit trees
I walked by the entrance of the [ETSIIT](http://etsiit.ugr.es) the
other day and noticed that the tree in front is a pomegranate tree, of
which there are a lot in Granada, probably second only to orange
trees. So I guess it spurred me to try this thing.

At the same time, I am no longer including any picture, mine or
otherwise, here. So I guess this closes #3.

I'll leave #2 for discussion. Maybe I should have two versions, the
printed version with pictures, or simply one with pictures and one
without. It's simply a matter of working with #15 and maybe #13.